feat: add entry
