+++ packages/expect/rollup.config.js
@@ -9,6 +9,7 @@ const external = [
   ...Object.keys(pkg.dependencies || {}),
   ...Object.keys(pkg.peerDependencies || {}),
   '@vitest/utils/diff',
+  '@vitest/utils/error',
 ]
 
 const plugins = [
