+++ packages/runner/rollup.config.js
@@ -9,7 +9,7 @@ const external = [
   ...builtinModules,
   ...Object.keys(pkg.dependencies || {}),
   ...Object.keys(pkg.peerDependencies || {}),
-  '@vitest/utils/diff',
+  '@vitest/utils/error',
 ]
 
 const entries = {
