+++ packages/snapshot/rollup.config.js
@@ -3,7 +3,6 @@ import esbuild from 'rollup-plugin-esbuild'
 import nodeResolve from '@rollup/plugin-node-resolve'
 import dts from 'rollup-plugin-dts'
 import commonjs from '@rollup/plugin-commonjs'
-import json from '@rollup/plugin-json'
 import { defineConfig } from 'rollup'
 import pkg from './package.json' assert { type: 'json' }
 
@@ -24,7 +23,6 @@ const plugins = [
     preferBuiltins: true,
   }),
   commonjs(),
-  json(),
   esbuild({
     target: 'node14',
   }),
