+++ packages/utils/src/index.ts
@@ -6,5 +6,5 @@ export * from './random'
 export * from './display'
 export * from './constants'
 export * from './colors'
-export * from './errors'
+export * from './base'
 export * from './source-map'
